feat: removed st40_utils monolith and moved helpers to fidasim utils
